Make AmqpDescribed members public per shipped API
- Offset/Length: internal int -> public long
- DecodeDescriptor: private static -> public static
- GetValueEncodeSize/EncodeValue/DecodeValue: internal virtual -> public virtual
- Update overrides in AmqpValue, Data, DescribedMap, DescribedList
- Cast Offset/Length to int in AmqpMessage.WriteBytes call

+++ src/AmqpMessage.cs
@@ -884,7 +884,7 @@ namespace Microsoft.Azure.Amqp
                 {
                     if (source != null)
                     {
-                        AmqpBitConverter.WriteBytes(buffer, source.Buffer, section.Offset, section.Length);
+                        AmqpBitConverter.WriteBytes(buffer, source.Buffer, (int)section.Offset, (int)section.Length);
                     }
                     else
                     {
